+++ eva/src/components/dashBoardInside/dashSingle.vue
@@ -169,6 +169,25 @@ export default {
         }
         flag = -1;
       });
+
+      events = this.$store.getters.getEvents({idDash: this.idDash, event: 'onValueCompare', element: this.id});
+
+      let treshold, color, value;
+
+      events.forEach( (item) => { 
+        treshold = item.treshold.replace('[','').replace(']','').split(',');
+        color = item.color.replace('[','').replace(']','').split(',');
+        treshold.forEach( (item,i) => {
+          if (number > Number(item)) {
+            value = color[i];
+          }
+        })
+        if (value) {
+          this.$store.commit('letEventSet', {events: [{target: item.target,prop: item.prop[0],value: value}], idDash: this.idDash,  });
+        } else {
+          this.$store.commit('letEventSet', {events: [{target: item.target,prop: item.prop[0],value: ''}], idDash: this.idDash,  });
+        }
+      });
             
     },
     setClick: function() {