+++ CHANGELOG.md
@@ -32,6 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - check permission for dashboard, main dash page, dash setting
 
+- add event onValueCopare
+
 ## [1.44.26] [newSetting]
 ### Adde